pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.50.0",
    "google-generativeai>=0.8.5",
    "openai>=1.76.0",
    "openpyxl>=3.1.5",
    "plotly>=6.0.1",
    "streamlit>=1.44.1",
]
